requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
decorator==4.4.2
Django==4.2.5
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.12.4
google-api-core==2.12.0
google-auth==2.23.2
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
Pillow==10.0.1
proglog==0.1.10
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.7.0
sqlparse==0.4.4
stomp.py==8.1.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
websocket-client==1.6.3
